Removed old XML file implementation
