+++ src/Workspace/Files/XxxxFile.ts
@@ -1,10 +1,7 @@
+import * as vscode from 'vscode';
 import { Uri } from 'vscode';
 import { logger } from '../../Tools/Logger';
-import { getChildElements } from '../../Tools/XmlDom';
 import { AsXmlFile } from './AsXmlFile';
-import { Element as XmlElement } from '@oozcitak/dom/lib/dom/interfaces';
-import { anyToBoolOrUndefined } from '../../Tools/Helpers';
-import { pathBasename, pathDirname, pathJoin, pathResolve, winPathToPosixPath } from '../../Tools/UriTools';
 
 /**
  * XXXXX file representation (*.xxxx)
@@ -19,9 +16,9 @@ export class XxxxFile extends AsXmlFile {
     public static async createFromPath(filePath: Uri): Promise<XxxxFile | undefined> {
         // Create and initialize object
         try {
-            const xmlFile = new XxxxFile(filePath);
-            await xmlFile._initialize();
-            return xmlFile;
+            const textDoc = await vscode.workspace.openTextDocument(filePath);
+            const fileContent = textDoc.getText();
+            return new XxxxFile(filePath, fileContent);
         } catch (error) {
             if (error instanceof Error) {
                 logger.error(`Failed to read XXXX file from path '${filePath.fsPath}': ${error.message}`); //TODO uri log #33
@@ -33,31 +30,18 @@ export class XxxxFile extends AsXmlFile {
         }
     }
 
-    /** Object is not ready to use after constructor due to async operations,
-     * _initialize() has to be called for the object to be ready to use! */
-    protected constructor(filePath: Uri) {
-        super(filePath);
-        // other properties rely on async and will be initialized in #initialize()
+    /** TODO doc */
+    protected constructor(filePath: Uri, fileContent: string) {
+        super(filePath, fileContent);
+        // initialize other properties here!
+        this.#xxxxx = '42';
     }
-
-    /**
-     * Async operations to finalize object construction
-     * @throws If a required initialization process failed
-     */
-    protected async _initialize(): Promise<void> {
-        await super._initialize();
-        //TODO
-        // init done
-        this.#isInitialized = true;
-    }
-    #isInitialized = false;
 
     /** XXXXXX */
     public get xxxxx(): string {
-        if (!this.#isInitialized || !this.#xxxxx) { throw new Error(`Use of not initialized ${XxxxFile.name} object`); }
         return this.#xxxxx;
     }
-    #xxxxx: string | undefined;
+    #xxxxx: string;
 
     /** toJSON required as getter properties are not shown in JSON.stringify() otherwise */
     public toJSON(): any {
